+++ packages/core/src/core/createLogController.js
@@ -10,18 +10,13 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { queryString, stringToBoolean } from "../utils/index.js";
-import debugQueryParam from "../constants/debugQueryParam.js";
-
 export default ({
   console,
-  locationSearch,
   createLogger,
   instanceName,
   createNamespacedStorage,
   getMonitors,
 }) => {
-  const parsedQueryString = queryString.parse(locationSearch);
   const storage = createNamespacedStorage(`instance.${instanceName}.`);
 
   const debugSessionValue = storage.session.getItem("debug");
@@ -40,12 +35,6 @@ export default ({
       debugEnabledWritableFromConfig = false;
     }
   };
-
-  if (parsedQueryString[debugQueryParam] !== undefined) {
-    setDebugEnabled(stringToBoolean(parsedQueryString[debugQueryParam]), {
-      fromConfig: false,
-    });
-  }
 
   return {
     setDebugEnabled,