+++ sandboxes/browser/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @adobe/alloy-sandbox-browser
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`543a0e9`](https://github.com/adobe/alloy/commit/543a0e9f7ddbcd0a79885dfba8e0539a8bcf4cf4), [`676b653`](https://github.com/adobe/alloy/commit/676b6535c7a17ef5879d2abff539ba721fc61bc7), [`44042df`](https://github.com/adobe/alloy/commit/44042df84fd4c003567e64aa3d03f6c94c33e094), [`601b4a8`](https://github.com/adobe/alloy/commit/601b4a8dcb964880fe7da518484a301891cabced)]:
+  - @adobe/alloy@2.34.0
+  - @adobe/alloy-core@1.2.0
+
+## 0.1.4-beta.2
+
+### Patch Changes
+
+- Updated dependencies [[`543a0e9`](https://github.com/adobe/alloy/commit/543a0e9f7ddbcd0a79885dfba8e0539a8bcf4cf4)]:
+  - @adobe/alloy@2.34.0-beta.2
+  - @adobe/alloy-core@1.2.0-beta.2
+
+## 0.1.4-beta.1
+
+### Patch Changes
+
+- Updated dependencies [[`44042df`](https://github.com/adobe/alloy/commit/44042df84fd4c003567e64aa3d03f6c94c33e094), [`601b4a8`](https://github.com/adobe/alloy/commit/601b4a8dcb964880fe7da518484a301891cabced)]:
+  - @adobe/alloy-core@1.2.0-beta.1
+  - @adobe/alloy@2.34.0-beta.1
+
 ## 0.1.4-beta.0
 
 ### Patch Changes